Extracted re-usable parts of the consensus caller. Changed the way consensusReadLength is calculated so that we don't try and generate consensus after we drop below minReads if we have reads of different lengths.
